removed comments

+++ cpp/oneapi/dal/algo/triangle_counting/backend/cpu/relabel_kernels.hpp
@@ -37,7 +37,7 @@ void sort_ids_by_degree(const std::int32_t* degrees,
         degree_id_pairs[n] = std::make_pair(degrees[n], (size_t)n);
     });
     dal::detail::parallel_sort(degree_id_pairs, degree_id_pairs + vertex_count);
-    dal::detail::threader_for(vertex_count / 2, vertex_count / 2, [&](std::int32_t i) {
+    dal::detail::threader_for(vertex_count / 2, vertex_count / 2, [&](std::int64_t i) {
         std::swap(degree_id_pairs[i], degree_id_pairs[vertex_count - i - 1]);
     });
 }
